refactor: move reuseable function
